return task id if user is admin
To make life easier for admins issuing tasks from the CLI, this
will return the task id if you requested the task as admin.

Adding a basic util to help with that.

Updating the TaskView approve function to not return a Response
so it can instead be built in the post function. Needed to update
the response dict with the task id, and also to make it a little
more consistent with process_actions which does not return Response.

Change-Id: Ieed96b165083d19c1bdf09621939716f8c79c5a6