Replaced find_region with get_region
Due to the weird way that keystone controls the ids for regions
find_region did not work, and was equivalent to get_region.

The id for a region is the same as it's name, which is different
to almost all other objects in keystone.

Find region has been removed and in the one place it was being
used it has been replaced with get_region.

Change-Id: Iadc361c6c15f1511eb8b6fef93e09f7d2503f2db