Changes to fake openstack client and caches
Due to more of stacktask interacting with the clients querying
about usage and deletion, the fake openstack clients have to
be updated.

Neutron cache top level key is now a region name, with the
second level being a project id.

A number of additional functions have been added to fake nova
client and fake neutron client to allow the user to query usage.

Change-Id: I3859f298b3530abaf2dc59f61d57e464f8a8a7f4