Added modify_dict_settings test decorator
This decorator allows for modification of test settings that are
stored as dictionaries for that tests only without needing to
override the whole dictionary. It acts in a similar way to django's
in built override_settings, and modify_settings.

To be applied to a class the class must subclass StacktaskTestCase
or StacktaskAPITestCase. In those two classes settings can also
be modified using:

with self.modify_dict_settings(...):
	# code

Example Usage:
@modify_dict_settings(ROLES_MAPPING=[
		    {'key_list': ['project_mod'],
		    'operation': 'remove',
		    'value': 'heat_stack_owner'},
		    {'key_list': ['project_admin'],
		    'operation': 'append',
		    'value': 'heat_stack_owner'},
		    ])
or
@modify_dict_settings(PROJECT_QUOTA_SIZES={
		    'key_list': ['small', 'nova', 'instances'],
		    'operations': 'override',
		    'value': 11
		    })

Available operations:
Standard operations:
- 'override': Either overrides or adds the value to the dictionary.
- 'delete': Removes the value from the dictionary.

List operations:
List operations expect that the accessed value in the dictionary
is a list.
- 'append': Add the specified values to the end of the list
- 'prepend': Add the specifed values to the start of the list
- 'remove': Remove the specified values from the list

Change-Id: I575c17891d14418c335b2fb8426830e6e31be515