Additional tests for USERNAME_IS_EMAIL=False
Note that while all of the base stacktask codebase works with this
setting, not all of the client and horizon will work with it.

This is due to when the settings change a number of actions now
also require a username field to be filled in, and there isn't a
way for the clients to determine whether or not this is the case.

Change-Id: Ie91f7634b3686a1535af1c3344be217db5a80d15