Additional Emails - Additional Action
Currently set up as an additional action added to the task in the
configuration file. At each stage (corresponding to the current email
section lables) a template and subject can be specified detailing
the email to be sent. This will be sent to the users email address
or otherwise an override email address set from the task.

In the configuration sending to the users email address can be
turned off with the line
   email_current_user: false

Additionally an email can be sent out to a group of roles within
a project using:
  email_roles:
      - project_admin

Or to a number of specific emails:
  email_additional_addresses:
      - admin@example.org

Or to an address specified in the task cache
  email_in_task_cache: true

(Cache key "additional_emails")
Change-Id: I6d454bdfefb7549322fea6cf0c91fac76b5aa89a