Added UpdateProjectQuota
Accessable at v1/openstack/quotas/.

A GET request returns json specifiing quota size, the current quota,
the current usage, and some details of any currently active quota
change tasks.

A POST request will update the quota for the project to a given
size. The data must contain a JSON dict with 'size' in it,
which will be the name of one of the pre-defined sizes.
Optionally regions (a list of region names) can be specified
which will restrict the update operation to those regions.

Change-Id: I907664f79f6eef0b5239139999cc7a28d246e446