Don't catch exceptions in quota manager
For these exceptions it is probably safer to actually
raise them and throw a 500. That way a notification
goes out, and we deal with it and potentially rerun the
actions.

The responsibility is on the deployer to correctly configure
the QUOTA_SERVICES setting, and eventually we will also add
checking against the catalog for service types in a given
region.

Change-Id: I3fa6013cb3dd02ae40109a10c79211e55cb73ef8