Add documentation for the project history and guide lines
Because of how vague the scope of Adjutant is (for a reason)
we need a way to define what actually fits in a sensible way.
The new guide lines are part of that, and we're also now
adding a section to our docs about the project history to
explain why the scope is so vague and how it got there.
Plus a section on project features, to help define what it
can do.

Change-Id: I7d05c1f43570843171a4144ece2228efc3227cce