Temporary fix for user enumeration via password reset API
There is a noticable delay when providing a valid username
to the password reset API. Ideally we want to fix this by handling
request in an async fashion, but that will likely have to wait
until we have moved to the planned worker/api model.

This just makes the API always take at least 3 seconds.

Change-Id: I82d46e9d64c65930dbe7d8821941ee9173431d56