Fixed typos in nova and neutron quota output
The Nova and Neutron quota helper objects output the quota data with
the word 'secuirty' instead of 'security' in the key name. Adjutant-ui,
nova and neutron (among others) look for the keys with the correct
spelling. No other instance of the typo appears to exist in other
modules.

APIImpact: some keys returned by Adjutant are now correctly spelled
Change-Id: Idf2953027765119d70720592ccc1187472854fad