Allow default networks in multiple regions
This adds the optional `create_in_regions` and `create_in_all_regions`
parameters to the `NewDefaultNetworkAction` and the
`NewProjectDefaultNetworkAction` actions.

When `create_in_regions` is set, the action will ignore the default
region set when creating a project, and instead create default networks
and routers for each region set in the `create_in_regions` list.
When `create_in_regions` is unset (or set to an empty list),
the actions will default to the behaviour dictated by
`create_in_all_regions`.

When `create_in_all_regions` is set to `False` (the default),
a default network is created in the default region only, as it
always has. When it is set to `True`, it will instead create
default networks in *all* active cloud regions,
as discovered by Adjutant.

Improve error handling in `NewDefaultNetworkAction` to give more
detail in the task log when resource creation fails.

Fix bugs in the fake Neutron client in the unit tests where created
resources were hard-coded to be added to the `RegionOne` region.

Change-Id: I0e0885dda449b22788bd4c9ae72b874d8ce661b5