Replace pkg_resources with importlib, remove debug docs.
As of Python 3.8, importlib.metadata is the provisional
replacement for pkg_resources which depended on setuptools.

In python 3.10 this is no longer provisional and supersedes
pkg_resources with the same functionality.

Also remove rest_framework_swagger which provides the `/docs/`
endpoint temporarily. The library is end of life and should be
replaced with drf-yasg or similar.

Ref: https://importlib-metadata.readthedocs.io/en/latest/migration.html
Ref: https://docs.python.org/3/library/importlib.metadata.html

Change-Id: I102e94c8172f6ea660814dd8fcd44ed8239b2e67