Fix flake8 F824 errors
A new flake8 test F824 was introduced in flake8 7.2.0 on 2025-03-30.

This errors if a global var is defined but never assigned to, as it
is not required to declare global when accessing variables outside
the current scope, only setting.

https://flake8.pycqa.org/en/latest/user/error-codes.html

Change-Id: I706568739ec4e5db50e2ef222758592851ed4496
Signed-off-by: Dale Smith <dale@catalystcloud.nz>
(cherry picked from commit 75d361177bb2e28e1fb0af3846c97f456d1aca4d)